修改数据库字段类型

+++ imking-web/src/main/java/io/imking/reward/beans/dto/RwApplyDetailDTO.java
@@ -26,5 +26,5 @@ public class RwApplyDetailDTO {
 
     private String askContent;
     
-    private BigDecimal taskAmount;
+    private Integer taskAmount;
 }
